Implement command class interfaces
