+++ src/main/java/commands/GuildPrefixManager.java
@@ -1,0 +1,12 @@
+package commands;
+
+// TODO: implement class
+public class GuildPrefixManager {
+    public static GuildPrefixManager getInstance() {
+        return null;
+    }
+
+    public String getPrefix(long guildId) {
+        return null;
+    }
+}
