+++ src/main/java/commands/newcommand/ISubCommand.java
@@ -1,0 +1,5 @@
+package commands.newcommand;
+
+public interface ISubCommand {
+    ParentCommand getParentCommand();
+}
