+++ src/main/java/commands/newcommand/ParentCommand.java
@@ -2,12 +2,6 @@ package commands.newcommand;
 
 import commands.newcommand.extensions.SlashCommand;
 import commands.newcommand.extensions.TextCommand;
-import net.dv8tion.jda.api.EmbedBuilder;
-import net.dv8tion.jda.api.entities.User;
-import net.dv8tion.jda.api.events.interaction.SlashCommandEvent;
-import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
-import net.dv8tion.jda.api.interactions.commands.build.CommandData;
-import utility.EmbedUtils;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -15,19 +9,18 @@ import java.util.List;
 public abstract class ParentCommand extends NewCommand implements IParentCommand {
     private List<SubCommand> subCommands = new ArrayList<>();
 
-    public CommandData getCommandData() {
-        throw new UnsupportedOperationException("This command does not have slash commands.");
-    }
-
+    @Override
     public void addSubCommand(SubCommand subCommand) {
         subCommand.assignParentCommand(this);
         this.subCommands.add(subCommand);
     }
 
+    @Override
     public List<SubCommand> getSubCommands() {
         return new ArrayList<>(subCommands);
     }
 
+    @Override
     public String getSubCommandsText(String prefix) {
         StringBuilder subCommandsText = new StringBuilder();
         for (SubCommand subCommand : getSubCommands()) {
